Restaurant 제작  1. Controller 및 서비스 수정 중  - owner와 restaurant ownerId의 일치 여부 확인  - 이미 등록된 점포는 추가되지 않음. 비즈니스 넘버를 통한 확인  2. 데이터 값이 제대로 저장이 안 되고 있어 해당 부분 수정이 필요.

+++ src/main/java/com/zerobase/babbook/service/RestaurantService.java
@@ -1,9 +1,10 @@
 package com.zerobase.babbook.service;
 
+import static com.zerobase.babbook.exception.ErrorCode.ALREADY_REGISTER_RESTAURANT;
+import static com.zerobase.babbook.exception.ErrorCode.DO_NOT_CORRECT_ACCESS;
 import static com.zerobase.babbook.exception.ErrorCode.NOT_FOUND_OWNER;
 import static com.zerobase.babbook.exception.ErrorCode.NOT_FOUND_RESTAURANT;
 
-import com.zerobase.babbook.domain.dto.RestaurantDto;
 import com.zerobase.babbook.domain.entity.Owner;
 import com.zerobase.babbook.domain.entity.Restaurant;
 import com.zerobase.babbook.domain.form.RestaurantForm;
@@ -11,7 +12,6 @@ import com.zerobase.babbook.domain.reprository.OwnerRepository;
 import com.zerobase.babbook.domain.reprository.RestaurantRepository;
 import com.zerobase.babbook.exception.CustomException;
 import java.util.List;
-import java.util.stream.Collectors;
 import javax.transaction.Transactional;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
@@ -22,36 +22,64 @@ public class RestaurantService {
     private final RestaurantRepository restaurantRepository;
     private final OwnerRepository ownerRepository;
     //레스토랑 추가
-    @Transactional
-    public Restaurant addRestaurant(RestaurantForm form,Owner owner ){
-        Owner o = ownerRepository.findById(owner.getId())
+
+    public String addRestaurant(RestaurantForm form, Owner owner){
+        ownerRepository.findById(owner.getId())
             .orElseThrow(() -> new CustomException(NOT_FOUND_OWNER));
-        //이미 있는 경우, 에러를 반환하고 싶음.
-        return restaurantRepository.save(Restaurant.of(form,owner));
+
+        if (isBusinessNumberExist(form.getBusinessNumber())) {
+            throw new CustomException(ALREADY_REGISTER_RESTAURANT);
+    }else {
+        Restaurant restaurant = add(form, owner.getId());
+        return "식당 등록에 성공하였습니다.";
     }
+}
+    private Restaurant add(RestaurantForm form,Long ownerId) {
+        return restaurantRepository.save(Restaurant.of(form,ownerId));
+    }
+    private boolean isBusinessNumberExist(String businessNumber) {
+        return restaurantRepository.findByBusinessNumber(businessNumber).isPresent();
+    }
+
     //레스토랑 업데이트
-    @Transactional
-    public Restaurant updateRestaurant(RestaurantForm form, Owner owner) {
-        Owner o = ownerRepository.findById(owner.getId())
+
+    public String updateRestaurant(RestaurantForm form, Owner owner) {
+        Restaurant restaurant = restaurantRepository
+            .findByBusinessNumber(form.getBusinessNumber())
+            .orElseThrow(() -> new CustomException(NOT_FOUND_RESTAURANT));
+        Restaurant check = restaurantRepository.findByOwnerId(
+                owner.getId())
             .orElseThrow(() -> new CustomException(NOT_FOUND_OWNER));
-        return restaurantRepository.save(Restaurant.of(form,owner));
+        if(!restaurant.equals(check)){//불일치
+            throw new CustomException(DO_NOT_CORRECT_ACCESS);
+        }else{
+            restaurantRepository.save(Restaurant.of(form,owner.getId()));
+            return "정상적으로 수정 되었습니다.";
+        }
     }
     //레스토랑 삭제
-    @Transactional
-    public void delete(Long restaurantId, Owner owner) {
-        Owner o = ownerRepository.findById(owner.getId())
-            .orElseThrow(() -> new CustomException(NOT_FOUND_OWNER));
+
+    public String delete(Long restaurantId, Owner owner) {
         Restaurant restaurant = restaurantRepository
             .findById(restaurantId)
-            .orElseThrow(()-> new CustomException(NOT_FOUND_RESTAURANT));
-         restaurantRepository.deleteById(restaurantId);
+            .orElseThrow(() -> new CustomException(NOT_FOUND_RESTAURANT));
+        Restaurant check = restaurantRepository.findByOwnerId(
+            owner.getId())
+            .orElseThrow(() -> new CustomException(NOT_FOUND_OWNER));
+        String name = restaurant.getName();
+       if(!restaurant.equals(check)){//불일치
+           throw new CustomException(DO_NOT_CORRECT_ACCESS);
+       }else{
+            restaurantRepository.deleteById(restaurantId);
+           return name +" 식당이 정상적으로 삭제 되었습니다.";
+       }
     }
     //간단하게 보이는 곳(전체 페이지 처리 필요.) 리스트//페이지 처리 필요. 별도로 만들기.
     public List<Restaurant> restaurantList() {//너무 많은 정보가 담기는 거 같긴 한데...
         return restaurantRepository.findAll();
     }
     //상세하게 보이게 하는 곳.(특정 데이터 전체 확인)
-    @Transactional
+
     public Restaurant restaurantDetail(Long restaurantId){
         return restaurantRepository.findById(restaurantId)
             .orElseThrow(() -> new CustomException(NOT_FOUND_RESTAURANT));
